Modify for Live Feed notification logic.

+++ src/main/java/sg/edu/nus/iss/voucher/feed/workflow/aws/service/SQSQueueHandler.java
@@ -7,6 +7,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
+import org.springframework.context.annotation.Profile;
 import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
@@ -18,27 +19,28 @@ import com.amazonaws.services.sqs.model.*;
 
 import sg.edu.nus.iss.voucher.feed.workflow.utility.GeneralUtility;
 
+@Profile("!test")
 @Service
 public class SQSQueueHandler {
 
 	@Value("${aws.sqs.queue.feed.name}")
-	private String queueName;
+	String queueName;
 
 	@Value("${aws.sns.feed.topic.arn}")
-	private String topicArn;
+	String topicArn;
 
 	@Autowired
 	private AmazonSQS amazonSQS;
 
 	@Autowired
 	private AmazonSNS amazonSNS;
-	
+
 	@Autowired
 	private SNSSubscriptionService snsSubscriptionService;
 
 	private static final Logger logger = LoggerFactory.getLogger(SQSQueueHandler.class);
 
-	@EventListener(ApplicationReadyEvent.class) // Runs when the application is ready
+	@EventListener(ApplicationReadyEvent.class)
 	public void createQueueIfNotExists() {
 		String queueUrl = "";
 		try {
@@ -71,36 +73,47 @@ public class SQSQueueHandler {
 	public void consumeMessages() {
 
 		try {
+			if (doesQueueExist(queueName)) {
+				String queueUrl = amazonSQS.getQueueUrl(queueName).getQueueUrl();
 
-			String queueUrl = amazonSQS.getQueueUrl(queueName).getQueueUrl();
+				logger.info("consumeMessages queueUrl:" + queueUrl);
 
-			logger.info("consumeMessages queueUrl:" + queueUrl);
+				ReceiveMessageRequest receiveMessageRequest = new ReceiveMessageRequest(queueUrl)
+						.withMaxNumberOfMessages(10).withWaitTimeSeconds(5);
 
-			ReceiveMessageRequest receiveMessageRequest = new ReceiveMessageRequest(queueUrl)
-					.withMaxNumberOfMessages(10).withWaitTimeSeconds(5);
+				List<Message> messages = amazonSQS.receiveMessage(receiveMessageRequest).getMessages();
 
-			List<Message> messages = amazonSQS.receiveMessage(receiveMessageRequest).getMessages();
+				for (Message message : messages) {
 
-			for (Message message : messages) {
+					logger.info("Received Message: " + message.getBody());
 
-				logger.info("Received Message: " + message.getBody());
+					processFeedData(message.getBody());
 
-				processFeedData(message.getBody());
+					amazonSQS.deleteMessage(new DeleteMessageRequest(queueUrl, message.getReceiptHandle()));
 
-				amazonSQS.deleteMessage(new DeleteMessageRequest(queueUrl, message.getReceiptHandle()));
-				
-				logger.info("Message deleted successfully.");
-				
+					logger.info("Message deleted successfully.");
+
+				}
 			}
 		} catch (Exception e) {
 			logger.error("Error consuming messages: " + e.getMessage());
 		}
 	}
 
-	private void processFeedData(String snsMessage) {
+	public boolean doesQueueExist(String queueName) {
+        // Get all queues in the account
+        List<String> queueUrls = amazonSQS.listQueues(new ListQueuesRequest()).getQueueUrls();
+
+        // Check if any of the URLs end with the desired queue name
+        return queueUrls.stream().anyMatch(url -> url.endsWith("/" + queueName));
+    }
+
+	public void processFeedData(String snsMessage) {
 
 		logger.info("Processing feed data: " + snsMessage);
-		
-		String retMsg= snsSubscriptionService.processNotification(snsMessage);
+
+		String retMsg = snsSubscriptionService.processNotification(snsMessage);
+
+		logger.info("processNotification: " + retMsg);
 	}
 }
